Update helpText for CommandAirbridge.java
Updated help text to clarify parameters.

Changed Max length to max distance to travel because if the user is standing on an airbridge and runs /minebot airbridge 7 the bot will walk 7 blocks and stop. This may or may not be intended. If it is a bug we should change it to walk to the end of the current bridge before building 7 blocks.